+++ src/components/navbar/title.tsx
@@ -17,15 +17,21 @@ export default function Title() {
     state.list,
   ]);
 
-  const [openAIKey, azureKey, env] = useOpenAIStore((state) => [
+  const [openAIKey, azureKey, openRouterKey, env] = useOpenAIStore((state) => [
     state.openai.apiKey,
     state.azure.apiKey,
+    state.openRouter.apiKey,
     state.env,
   ]);
 
   const apiKey = React.useMemo(
-    () => openAIKey || azureKey || env.OPENAI_API_KEY || env.AZURE_API_KEY,
-    [openAIKey, azureKey, env]
+    () =>
+      openAIKey ||
+      azureKey ||
+      openRouterKey ||
+      env.OPENAI_API_KEY ||
+      env.AZURE_API_KEY,
+    [openAIKey, azureKey, openRouterKey, env]
   );
 
   const { channel_name, channel_cost } = React.useMemo(() => {