+++ src/hooks/useLLM/index.tsx
@@ -9,11 +9,6 @@ export const PREMIUM_MODELS = [
   "gpt-4-0613",
   "gpt-4-32k",
   "gpt-4-32k-0613",
-  "Claude Instant v1",
-  "Claude v2",
-  "PaLM 2 Bison",
-  "Llama v2 13B",
-  "Llama v2 70B",
 ];
 
 export const useLLMStore = createWithEqualityFn<LLMStore>(
@@ -86,7 +81,6 @@ export const useLLMStore = createWithEqualityFn<LLMStore>(
           icon: (
             <Image src="/claude.webp" alt="Claude" width={16} height={16} />
           ),
-          premium: true,
         },
         {
           label: "Claude v2",
@@ -94,25 +88,21 @@ export const useLLMStore = createWithEqualityFn<LLMStore>(
           icon: (
             <Image src="/claude.webp" alt="Claude" width={16} height={16} />
           ),
-          premium: true,
         },
         {
           label: "PaLM 2 Bison",
           value: "google/palm-2-chat-bison",
           icon: <Image src="/palm.webp" alt="PaLM" width={16} height={16} />,
-          premium: true,
         },
         {
           label: "Llama v2 13B",
           value: "meta-llama/llama-2-13b-chat",
           icon: <div>🦙</div>,
-          premium: true,
         },
         {
           label: "Llama v2 70B",
           value: "meta-llama/llama-2-70b-chat",
           icon: <div>🦙</div>,
-          premium: true,
         },
       ],
     },